Move utils into single files, fix logic, add comment about isNaN, use constants file, & add spacing after calls.
