+++ assets/js/googlesitekit/datastore/user/utils/constants.js
@@ -1,0 +1,2 @@
+export const INVALID_DATE_INSTANCE_ERROR = 'Date param must construct to a valid date instance or be a valid date instance itself.';
+export const INVALID_DATE_STRING_ERROR = 'Invalid dateString parameter, it must be a string.';
